edit

+++ backend/controllers/user.controller.js
@@ -35,12 +35,18 @@ module.exports.deleteAccount = async(req, res) => {
     if(!user){
         return res.status(400).json({ message: 'User not found' });
     }
+    const pswd = req.body.password;
+    const isPasswordValid = await bcrypt.compare(pswd, user.password);
+    if(!isPasswordValid){
+        return res.status(400).json({ message: 'Invalid password' });
+    }
     await UserModel.deleteOne({login: req.body.login});
     res.status(200).json({ message: 'Account deleted' });
 };
 
 module.exports.editUser = async(req, res) => {
     const user = await UserModel.findOne({login: req.body.login});
+    console.log(user);
     if(!user){
         return res.status(400).json({ message: 'User not found' });
     }
